Add 'all people' tab
This adds an all people tab which displays everybody who is registered in the system.
This is very similar to the participants table, except that it also shows lecturers, is not filtered by year,
and displays different columns by default.
Note that a visible by default 'years this user participated in' column will be added by #203

Closes #88