Make the workshop template apply upon clicking save for the first time instead of first editor load
I'm not sure who designed it that way originally, but this seems more
logical. With the new workshop status buttons and workshops being directly
editable by admins, the template was getting applied as soon as the
workshop was accepted. This makes it possible to edit the template
between accepting the workshop and the lecturer acutally filling out
the page.