Make the CampParticipant object exist as soon as the user registers for workshops
Until now, this object was basically used as a "camp qualification result" object,
created only after the user was qualified. Now, the object will be created
immediately after the user registers for the first workshop. This is ensured
by the database model - WorkshopParticipant now links to the CampParticipant
object. This cleans up some of the funky who-participated-when code, and makes
it possible to store per-year-per-camp user information before they are
qualified, like the cover letter.