Fix empty tables in profile page
This makes the box appear only when the user was a participant,
since the lecturer data is in another place in this view.
This is how it was originally before I broke it.

Also added a message for people who expressed interest before
the program was available, to make the box not empty

Closes #584