Move CampParticipant count calculations into the model
These used to live in the view before we had the CampParticipant object,
because there was no better place to put them. It makes more sense for them
to be there now.

Also fixed some edge cases when disabling uploads/scoring after some results
are already in the system and added tests.